application variable properties : see QuizbeGlobals.kt

+++ src/main/kotlin/org/quizbe/config/QuizbeProperties.kt
@@ -6,6 +6,6 @@ import org.springframework.boot.context.properties.ConstructorBinding
 // not working (see QuizbeGlobals as temporary solution)
 @ConstructorBinding
 @ConfigurationProperties("quizbe")
-data class QuizbeProperties(var pwLifeTimeHours: Int) {
+data class QuizbeProperties(var pwLifeTimeHours: Long) {
 
 }
